refactor(dashboard): overhaul event management components for improved structure and performance
- Reorganized event management components to enhance readability and maintainability.
- Introduced new content components for analytics, attendees, details, promotions, payments, and tickets, streamlining the event management process.
- Implemented lazy loading and suspense for better loading state handling across various pages.
- Updated routing constants for clarity and consistency in navigation.
- Cleaned up imports and improved code organization throughout the dashboard.
