+++ src/app/(dashboard)/dashboard/events/[id]/payments/page.tsx
@@ -1,176 +1,19 @@
-import { Button } from '@/components/ui/button';
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card';
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/ui/table';
-import { Badge } from '@/components/ui/badge';
-import { Download, Filter } from 'lucide-react';
+"use client";
 
-const transactions = [
-  {
-    id: 'TX123',
-    date: '2024-03-15',
-    customer: 'Sarah Davis',
-    event: 'Tech Conference 2025',
-    amount: 299.00,
-    status: 'completed',
-  },
-  {
-    id: 'TX124',
-    date: '2024-03-14',
-    customer: 'Michael Johnson',
-    event: 'Music Festival',
-    amount: 150.00,
-    status: 'completed',
-  },
-  {
-    id: 'TX125',
-    date: '2024-03-14',
-    customer: 'Emily Brown',
-    event: 'Food & Wine Expo',
-    amount: 75.00,
-    status: 'pending',
-  },
-];
+import { Suspense, use } from "react";
 
-export default function PaymentsPage() {
+import { PaymentsContent } from "@/components/dashboard/event/payments/PaymentsContent";
+import Spinner from "@/components/ui/spinner";
+
+type Props = {
+  params: Promise<{ id: string }>;
+};
+
+export default function PaymentsPage({ params }: Props) {
+  const { id } = use(params);
   return (
-    <div className="space-y-6 animate-in">
-      <div className="flex justify-between items-center">
-        <div>
-          <h1 className="text-3xl font-bold">Payments</h1>
-          <p className="text-muted-foreground">
-            Manage your event payments and transactions
-          </p>
-        </div>
-        <div className="flex gap-4">
-          <Button variant="outline" className="gap-2">
-            <Filter className="h-4 w-4" />
-            Filter
-          </Button>
-          <Button variant="outline" className="gap-2">
-            <Download className="h-4 w-4" />
-            Export
-          </Button>
-        </div>
-      </div>
-
-      <div className="grid gap-6 md:grid-cols-4">
-        <Card>
-          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-            <CardTitle className="text-sm font-medium">
-              Total Revenue
-            </CardTitle>
-          </CardHeader>
-          <CardContent>
-            <div className="text-2xl font-bold">$45,231.89</div>
-            <p className="text-xs text-muted-foreground">
-              +20.1% from last month
-            </p>
-          </CardContent>
-        </Card>
-        <Card>
-          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-            <CardTitle className="text-sm font-medium">
-              Pending Payments
-            </CardTitle>
-          </CardHeader>
-          <CardContent>
-            <div className="text-2xl font-bold">$2,450.00</div>
-            <p className="text-xs text-muted-foreground">
-              5 transactions pending
-            </p>
-          </CardContent>
-        </Card>
-        <Card>
-          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-            <CardTitle className="text-sm font-medium">
-              Average Ticket Price
-            </CardTitle>
-          </CardHeader>
-          <CardContent>
-            <div className="text-2xl font-bold">$175.00</div>
-            <p className="text-xs text-muted-foreground">
-              Across all events
-            </p>
-          </CardContent>
-        </Card>
-        <Card>
-          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-            <CardTitle className="text-sm font-medium">
-              Refunds
-            </CardTitle>
-          </CardHeader>
-          <CardContent>
-            <div className="text-2xl font-bold">$350.00</div>
-            <p className="text-xs text-muted-foreground">
-              2 refunds this month
-            </p>
-          </CardContent>
-        </Card>
-      </div>
-
-      <Card>
-        <CardHeader>
-          <CardTitle>Recent Transactions</CardTitle>
-          <CardDescription>
-            A list of your recent transactions
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <div className="border rounded-lg">
-            <Table>
-              <TableHeader>
-                <TableRow>
-                  <TableHead>Transaction ID</TableHead>
-                  <TableHead>Date</TableHead>
-                  <TableHead>Customer</TableHead>
-                  <TableHead>Event</TableHead>
-                  <TableHead>Amount</TableHead>
-                  <TableHead>Status</TableHead>
-                  <TableHead>Actions</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {transactions.map((transaction) => (
-                  <TableRow key={transaction.id}>
-                    <TableCell className="font-medium">
-                      {transaction.id}
-                    </TableCell>
-                    <TableCell>{transaction.date}</TableCell>
-                    <TableCell>{transaction.customer}</TableCell>
-                    <TableCell>{transaction.event}</TableCell>
-                    <TableCell>${transaction.amount.toFixed(2)}</TableCell>
-                    <TableCell>
-                      <Badge
-                        variant={transaction.status === 'completed' ? 'default' : 'secondary'}
-                      >
-                        {transaction.status}
-                      </Badge>
-                    </TableCell>
-                    <TableCell>
-                      <Button variant="ghost" size="sm">
-                        View Details
-                      </Button>
-                    </TableCell>
-                  </TableRow>
-                ))}
-              </TableBody>
-            </Table>
-          </div>
-        </CardContent>
-      </Card>
-    </div>
+    <Suspense fallback={<Spinner />}>
+      <PaymentsContent eventId={id} />
+    </Suspense>
   );
 }
