+++ src/components/dashboard/LazyDashboardComponents.tsx
@@ -1,71 +1,87 @@
-import dynamic from 'next/dynamic';
+"use client";
+
 import { Skeleton } from "@/components/ui/skeleton";
+import dynamic from "next/dynamic";
 
 // Lazy load the heavy Overview component with Recharts (~100KB)
-export const OverviewContent = dynamic(() => import("@/components/dashboard/overview").then(mod => ({ default: mod.Overview })), {
-  loading: () => (
-    <div className="space-y-4">
-      <div className="space-y-2">
-        <Skeleton className="h-6 w-48" />
-        <Skeleton className="h-4 w-64" />
+export const OverviewContent = dynamic(
+  () =>
+    import("@/components/dashboard/overview").then((mod) => ({
+      default: mod.Overview,
+    })),
+  {
+    loading: () => (
+      <div className="space-y-4">
+        <div className="space-y-2">
+          <Skeleton className="h-6 w-48" />
+          <Skeleton className="h-4 w-64" />
+        </div>
+        <Skeleton className="h-[350px] w-full rounded-xl" />
       </div>
-      <Skeleton className="h-[350px] w-full rounded-xl" />
-    </div>
-  ),
-  ssr: false, // Charts don't need SSR
-});
+    ),
+  }
+);
 
 // Lazy load heavy form components
-export const EventForm = dynamic(() => import("@/components/forms/event-form"), {
-  loading: () => (
-    <div className="space-y-6">
-      <div className="space-y-2">
-        <Skeleton className="h-4 w-24" />
-        <Skeleton className="h-10 w-full" />
+export const EventForm = dynamic(
+  () => import("@/components/forms/event-form"),
+  {
+    loading: () => (
+      <div className="space-y-6">
+        <div className="space-y-2">
+          <Skeleton className="h-4 w-24" />
+          <Skeleton className="h-10 w-full" />
+        </div>
+        <div className="space-y-2">
+          <Skeleton className="h-4 w-32" />
+          <Skeleton className="h-32 w-full" />
+        </div>
+        <div className="flex gap-4">
+          <Skeleton className="h-10 w-24" />
+          <Skeleton className="h-10 w-24" />
+        </div>
       </div>
-      <div className="space-y-2">
-        <Skeleton className="h-4 w-32" />
-        <Skeleton className="h-32 w-full" />
-      </div>
-      <div className="flex gap-4">
-        <Skeleton className="h-10 w-24" />
-        <Skeleton className="h-10 w-24" />
-      </div>
-    </div>
-  ),
-});
+    ),
+  }
+);
 
-export const TicketForm = dynamic(() => import("@/components/forms/ticket-form"), {
-  loading: () => (
-    <div className="space-y-6">
-      <div className="space-y-2">
-        <Skeleton className="h-4 w-24" />
-        <Skeleton className="h-10 w-full" />
+export const TicketForm = dynamic(
+  () => import("@/components/forms/ticket-form"),
+  {
+    loading: () => (
+      <div className="space-y-6">
+        <div className="space-y-2">
+          <Skeleton className="h-4 w-24" />
+          <Skeleton className="h-10 w-full" />
+        </div>
+        <div className="space-y-2">
+          <Skeleton className="h-4 w-32" />
+          <Skeleton className="h-32 w-full" />
+        </div>
+        <div className="flex gap-4">
+          <Skeleton className="h-10 w-24" />
+          <Skeleton className="h-10 w-24" />
+        </div>
       </div>
-      <div className="space-y-2">
-        <Skeleton className="h-4 w-32" />
-        <Skeleton className="h-32 w-full" />
-      </div>
-      <div className="flex gap-4">
-        <Skeleton className="h-10 w-24" />
-        <Skeleton className="h-10 w-24" />
-      </div>
-    </div>
-  ),
-});
+    ),
+  }
+);
 
-export const TeamMembersForm = dynamic(() => import("@/components/forms/team-members-form"), {
-  loading: () => (
-    <div className="space-y-6">
-      <Skeleton className="h-8 w-48" />
-      <div className="grid gap-4">
-        <Skeleton className="h-16 w-full" />
-        <Skeleton className="h-16 w-full" />
-        <Skeleton className="h-16 w-full" />
+export const TeamMembersForm = dynamic(
+  () => import("@/components/forms/team-members-form"),
+  {
+    loading: () => (
+      <div className="space-y-6">
+        <Skeleton className="h-8 w-48" />
+        <div className="grid gap-4">
+          <Skeleton className="h-16 w-full" />
+          <Skeleton className="h-16 w-full" />
+          <Skeleton className="h-16 w-full" />
+        </div>
       </div>
-    </div>
-  ),
-});
+    ),
+  }
+);
 
 // Create skeleton components for dashboard sections
 export function StatsSkeleton() {
@@ -90,7 +106,10 @@ export function ActivitySkeleton() {
         <Skeleton className="h-4 w-48" />
       </div>
       {Array.from({ length: 3 }).map((_, i) => (
-        <div key={i} className="flex items-center space-x-4 p-4 border rounded-lg">
+        <div
+          key={i}
+          className="flex items-center space-x-4 p-4 border rounded-lg"
+        >
           <Skeleton className="h-10 w-10 rounded-full" />
           <div className="space-y-2 flex-1">
             <Skeleton className="h-4 w-32" />
@@ -106,7 +125,10 @@ export function CategoriesSkeleton() {
   return (
     <div className="grid gap-4 md:grid-cols-2">
       {Array.from({ length: 4 }).map((_, i) => (
-        <div key={i} className="flex items-center space-x-4 p-4 border rounded-lg">
+        <div
+          key={i}
+          className="flex items-center space-x-4 p-4 border rounded-lg"
+        >
           <Skeleton className="h-12 w-12 rounded-lg" />
           <div className="space-y-2 flex-1">
             <Skeleton className="h-5 w-24" />
@@ -126,7 +148,10 @@ export function ReferralSkeleton() {
         <Skeleton className="h-4 w-48" />
       </div>
       {Array.from({ length: 3 }).map((_, i) => (
-        <div key={i} className="flex items-center justify-between p-4 border rounded-lg">
+        <div
+          key={i}
+          className="flex items-center justify-between p-4 border rounded-lg"
+        >
           <div className="space-y-2">
             <Skeleton className="h-4 w-24" />
             <Skeleton className="h-3 w-32" />
@@ -136,4 +161,4 @@ export function ReferralSkeleton() {
       ))}
     </div>
   );
-} 
+}
