+++ src/components/dashboard/event/edit/EditEventContent.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import EventForm from "@/components/forms/event-form";
+
+export function EditEventContent({ eventId }: { eventId: string }) {
+  return (
+    <div className="min-h-screen bg-white dark:bg-zinc-900 text-gray-900 dark:text-gray-200 p-6">
+      <div className="max-w-6xl mx-auto">
+        <EventForm />
+      </div>
+    </div>
+  );
+}
