+++ src/components/forms/team-members-form.tsx
@@ -2,216 +2,195 @@
 
 import * as z from "zod";
 
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/ShadCard";
-import { useForm } from "react-hook-form";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormMessage,
-} from "@/components/ui/form";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
+import { Loader2, Plus, Trash2 } from "lucide-react";
+import { useEffect, useState } from "react";
 
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { useInviteTeamMember, useTeamMembers } from "@/hooks/useEvent";
-import { usePublishEvent } from "@/hooks/useEvent"; // Import the usePublish hook
-import { useParams, useSearchParams } from "next/navigation";
-// import { useTeamMembers } from "@/services/events-team/queries"; // Import the query hook
-import { zodResolver } from "@hookform/resolvers/zod";
+import { Label } from "../ui/label";
 import { TeamMemberRole } from "@/types/event";
-import { Users } from "lucide-react";
-
-// First, define the role type and schema
-const ROLES = {
-  EVENT_ORGANIZER: "EVENT_ORGANIZER",
-  EVENT_TEAM: "EVENT_TEAM",
-  SELLER: "SELLER",
-  CUSTOMER: "CUSTOMER",
-} as const;
-
-type Role = keyof typeof ROLES;
+import toast from "react-hot-toast";
+import { useEvent } from "@/hooks/useEvent";
+import { useSearchParams } from "next/navigation";
 
 const teamSchema = z.object({
   user_email: z.string().email("Please enter a valid email address"),
-  role: z.enum(["EVENT_ORGANIZER", "EVENT_TEAM", "SELLER", "CUSTOMER"], {
+  role: z.nativeEnum(TeamMemberRole, {
     required_error: "Please select a role",
   }),
 });
 
+interface TeamMemberFormData {
+  email: string;
+  role: TeamMemberRole;
+}
+
 export default function TeamMembersForm() {
-
-  const params = useParams();
-  const eventId = params.id as string;
-  // Initialize the form
-  const teamForm = useForm<z.infer<typeof teamSchema>>({
-    resolver: zodResolver(teamSchema),
-    defaultValues: {
-      user_email: "", // Adjusted to match the new schema
-      role: "EVENT_TEAM", // Default role
-    },
-  });
-
   const searchParams = useSearchParams();
-  // const eventId = searchParams.get("eventId");
+  const eventId = searchParams.get("id") as string;
 
-  // Fetch current team members
-  const { data: teamMembers, refetch } = useTeamMembers(eventId);
+  const [teamMembers, setTeamMembers] = useState<TeamMemberFormData[]>([
+    { email: "", role: TeamMemberRole.EVENT_TEAM },
+  ]);
+  const [isLoading, setIsLoading] = useState(false);
 
-  const inviteTeamMember = useInviteTeamMember(eventId); // Use the mutation to invite a team member
-  console.log({teamMembers})
-  const handleInvite = async () => {
-    const memberData = {
-      user_email: teamForm.getValues("user_email"),
-      role: teamForm.getValues("role") as TeamMemberRole,
-    }; // Get email and role from form
-    if (eventId)
-      await inviteTeamMember.mutateAsync(memberData);
-    console.log({ memberData });
-    refetch(); // Refetch team members after inviting
+  const { useTeamMembers, useInviteTeamMember, useRemoveTeamMember } =
+    useEvent();
+
+  const { data: teamMembersData, error: teamMembersError } =
+    useTeamMembers(eventId);
+  const inviteTeamMember = useInviteTeamMember(eventId);
+  const removeTeamMember = useRemoveTeamMember(eventId);
+
+  useEffect(() => {
+    if (teamMembersData?.result?.results) {
+      setTeamMembers(
+        teamMembersData.result.results.map((member) => ({
+          email: member.email,
+          role: member.role,
+        }))
+      );
+    }
+  }, [teamMembersData]);
+
+  const handleAddMember = () => {
+    setTeamMembers([
+      ...teamMembers,
+      { email: "", role: TeamMemberRole.EVENT_TEAM },
+    ]);
   };
 
-  // Initialize the publish mutation
-  const publishEvent = usePublishEvent(eventId); // Use the publish hook
+  const handleRemoveMember = async (index: number) => {
+    const member = teamMembersData?.result?.results[index];
+    if (member) {
+      try {
+        await removeTeamMember.mutateAsync(member.id);
+        toast.success("Team member removed successfully");
+      } catch (error) {
+        toast.error("Failed to remove team member");
+      }
+    }
+    const newMembers = teamMembers.filter((_, i) => i !== index);
+    setTeamMembers(newMembers);
+  };
 
-  const handlePublish = async () => {
+  const handleMemberChange = (
+    index: number,
+    field: keyof TeamMemberFormData,
+    value: string
+  ) => {
+    const newMembers = [...teamMembers];
+    newMembers[index][field] = value as TeamMemberRole;
+    setTeamMembers(newMembers);
+  };
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setIsLoading(true);
+
     try {
-      // Call the publish function (you may need to pass eventId or other data)
-      if (eventId) await publishEvent.mutateAsync(); // Adjust as necessary
-      // Optionally, handle success (e.g., show a success message)
-      console.log("Event published successfully!");
+      // Invite each new team member
+      const invitePromises = teamMembers.map((member) =>
+        inviteTeamMember.mutateAsync({
+          user_email: member.email,
+          role: member.role,
+        })
+      );
+
+      await Promise.all(invitePromises);
+      toast.success("Team members invited successfully");
     } catch (error) {
-      // Handle error (e.g., show an error message)
-      console.error("Error publishing event:", error);
+      toast.error("Failed to invite team members");
+    } finally {
+      setIsLoading(false);
     }
   };
 
+  if (teamMembersError) {
+    return (
+      <div className="text-red-500">
+        Error loading team members. Please try again.
+      </div>
+    );
+  }
+
   return (
-    <div className="space-y-8">
-      {/* Card to display current team members */}
-      <div className="space-y-6">
-        <div>
-          <h2 className="text-2xl font-bold text-foreground mb-2">
-            Invite Team Members
-          </h2>
-          <p className="text-muted-foreground">
-            Add team members to help manage your event
-          </p>
-        </div>
-        
-        {/* Form to invite a new team member */}
-        <Form {...teamForm}>
-          <form
-            onSubmit={teamForm.handleSubmit(handleInvite)}
-            className="flex flex-col gap-4 sm:flex-row sm:items-end sm:gap-4"
-          >
-            {/* Input for team member email */}
-            <FormField
-              control={teamForm.control}
-              name="user_email" // Updated to match the new schema
-              render={({ field }) => (
-                <FormItem className="flex-1">
-                  <FormControl>
-                    <Input
-                      {...field}
-                      type="email"
-                      placeholder="team@example.com"
-                      className="rounded-xl"
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            {/* Dropdown for role selection */}
-            <FormField
-              control={teamForm.control}
-              name="role" // Updated to match the new schema
-              render={({ field }) => (
-                <FormItem className="w-full sm:w-48">
-                  <Select
-                    onValueChange={field.onChange}
-                    value={field.value}
-                  >
-                    <SelectTrigger className="rounded-xl">
-                      <SelectValue placeholder="Select Role" />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="EVENT_ORGANIZER">
-                        Event Organizer
-                      </SelectItem>
-                      <SelectItem value="EVENT_TEAM">Event Team</SelectItem>
-                      <SelectItem value="SELLER">Seller</SelectItem>
-                      <SelectItem value="CUSTOMER">Customer</SelectItem>
-                    </SelectContent>
-                  </Select>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <Button
-              type="submit"
-              className="rounded-xl px-6 py-2 w-full sm:w-auto"
-            >
-              Invite Team Member
-            </Button>
-          </form>
-        </Form>
-      </div>
-      
-      {/* Current Team Members Section */}
+    <form onSubmit={handleSubmit} className="space-y-6">
       <div className="space-y-4">
-        <div className="border-t border-border pt-6">
-          <h3 className="text-xl font-bold text-foreground mb-4">Current Team Members</h3>
-          
-          {teamMembers?.result?.results && teamMembers.result.results.length > 0 ? (
-            <div className="overflow-hidden rounded-xl border border-border">
-              <table className="min-w-full divide-y divide-border">
-                <thead className="bg-muted/50">
-                  <tr>
-                    <th className="px-6 py-3 text-left text-sm font-semibold text-foreground uppercase tracking-wider">
-                      Email
-                    </th>
-                    <th className="px-6 py-3 text-left text-sm font-semibold text-foreground uppercase tracking-wider">
-                      Role
-                    </th>
-                  </tr>
-                </thead>
-                <tbody className="bg-background divide-y divide-border">
-                  {teamMembers.result.results.map((member) => (
-                    <tr key={member.id} className="hover:bg-muted/30 transition-colors">
-                      <td className="px-6 py-4 whitespace-nowrap text-sm text-foreground">
-                        {member.email}
-                      </td>
-                      <td className="px-6 py-4 whitespace-nowrap text-sm text-muted-foreground">
-                        {member.role}
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
+        {teamMembers.map((member, index) => (
+          <div
+            key={index}
+            className="grid grid-cols-1 md:grid-cols-2 gap-4 p-4 border rounded-lg"
+          >
+            <div className="space-y-2">
+              <Label htmlFor={`email-${index}`}>Email</Label>
+              <Input
+                id={`email-${index}`}
+                type="email"
+                value={member.email}
+                onChange={(e) =>
+                  handleMemberChange(index, "email", e.target.value)
+                }
+                placeholder="Enter email"
+                required
+              />
             </div>
-          ) : (
-            <div className="text-center py-12 rounded-xl border border-dashed border-border">
-              <Users className="mx-auto h-12 w-12 text-muted-foreground mb-4" />
-              <p className="text-muted-foreground">No team members yet</p>
-              <p className="text-sm text-muted-foreground">Invite your first team member to get started</p>
+            <div className="space-y-2">
+              <Label htmlFor={`role-${index}`}>Role</Label>
+              <select
+                id={`role-${index}`}
+                value={member.role}
+                onChange={(e) =>
+                  handleMemberChange(index, "role", e.target.value)
+                }
+                className="w-full p-2 border rounded-md"
+                required
+              >
+                {Object.values(TeamMemberRole).map((role) => (
+                  <option key={role} value={role}>
+                    {role.replace(/_/g, " ")}
+                  </option>
+                ))}
+              </select>
             </div>
-          )}
-        </div>
+            {index > 0 && (
+              <Button
+                type="button"
+                variant="destructive"
+                size="sm"
+                className="md:col-span-2"
+                onClick={() => handleRemoveMember(index)}
+              >
+                <Trash2 className="h-4 w-4 mr-2" />
+                Remove Member
+              </Button>
+            )}
+          </div>
+        ))}
       </div>
-    </div>
+
+      <div className="flex justify-between">
+        <Button
+          type="button"
+          variant="outline"
+          onClick={handleAddMember}
+          className="flex items-center"
+        >
+          <Plus className="h-4 w-4 mr-2" />
+          Add Team Member
+        </Button>
+
+        <Button type="submit" disabled={isLoading}>
+          {isLoading ? (
+            <>
+              <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+              Saving...
+            </>
+          ) : (
+            "Save Changes"
+          )}
+        </Button>
+      </div>
+    </form>
   );
 }
