+++ src/types/ticket.ts
@@ -118,17 +118,6 @@ export interface CreateTicketTypeRequest {
   sale_end_date: Date;
 }
 
-export interface TicketTypeUpdateRequest {
-  id: string;
-  name?: string;
-  description?: string;
-  price?: number;
-  quantity?: number;
-  is_active?: boolean;
-  sale_start_date?: Date;
-  sale_end_date?: Date;
-}
-
 export interface UpdateTicketTypeRequest
   extends Partial<CreateTicketTypeRequest> {}
